[util] Rename Recent Work components
