+++ src/components/ui/blocks/recent-works-blocks/RecentWorks.js
@@ -1,28 +1,8 @@
 import React from 'react';
-import Container from 'react-bootstrap/Container';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 
-import { workSet } from '../../../../constants/globalConstants';
-
-const RecentWorkBlock = () => {
-    return (
-        <>
-            <Container>
-                <Row className="justify-content-center align-content-center" style={{ height: '55vh' }}>
-                    <Col className="col-12">
-                        <h2 style={{ fontSize: '2rem', margin: '2rem auto' }}>
-                            Recent Works
-                        </h2>
-                    </Col>
-                    <RecentWroks workSet={workSet} />
-                </Row>
-            </Container>
-        </>
-    );
-}
-
-const RecentWroks = ({ workSet }) => {
+const RecentWorks = ({ workSet }) => {
     return Object.keys(workSet).map(work => {
         let techs = [];
         techs = workSet[work].techs.map(tech => {
@@ -51,4 +31,4 @@ const RecentWroks = ({ workSet }) => {
     });
 }
 
-export default RecentWorkBlock;
+export default RecentWorks;
